update actions and reducers folder to include signupActions.js and signupReducer.js; still refactoring redux in app
