+++ src/store/actions/index.js
@@ -1,2 +1,2 @@
-export * from './authActions';
+export * from './signupActions';
 export * from './poemActions';
